build: Add minimal configure.ac
When building with slibtool using the rlibtool symlink the build will
fail when it fails to find the generated libtool. This file is required
for rlibtool because it is used to determine if building shared or
static.

This will require these commands to build:

  autoreconf -fi
  ./configure
  make

Gentoo Bug: https://bugs.gentoo.org/778911